requirements.txt
jieba
flask
pydict-cedar